requirements.txt
gradio
newspaper3k
transformers
torch
pandas
langdetect
lxml-html-clean